Before:

```
ivan@P50:~/release/gensim$ time python -c "import gensim"
Using TensorFlow backend.

real	0m1,008s
user	0m0,924s
sys	0m0,556s
```

After:
```
ivan@P50:~/release/gensim$ time python -c "import gensim"

real	0m0,299s
user	0m0,332s
sys	0m0,400s
```