@gojomo 2017-10-06
I don't think the original complaint was just a 0.75s delay, so doesn't seem like the numbers here show the real loading cost. 

I believe the best fix would move the Keras-specific convenience method elsewhere entirely. (Also, while I believe the effective re-`import` every time `get_embedding_layer()` is called is 'mostly harmless', it doesn't seem ideal to me.)