@menshikh-iv 2017-10-06
@gojomo unfortunately, if we moved this method to another place - we will lose this function.
FYI - "real import" will be run only once, with the first call of `get_embedding_layer()`, next call will be used already imported `keras` (re-import will not happen).