@gojomo 2017-10-06
Why would moving it elsewhere lose the functionality? 

In my opinion it'd be fine to even just define inline in the two files that use it (one test file, one notebook). It's a 1-liner – really just a way to remind people how `syn0` can be supplied to the keras `Embedding` constructor.