@menshikh-iv 2017-10-06
@gojomo 

> Why would moving it elsewhere lose the functionality?

Because it becomes difficult to find it -> nobody will use this (for example, we move this to `gensim.utls`)

Or your propose removing this method, replace `get_embeding_layer()` everywhere to direct call (in tests and notebook), like

```python
    layer = Embedding(
        input_dim=kv.syn0.shape[0], output_dim=kv.syn0.shape[1],
        weights=[kv.syn0], trainable=False
    )
```