@gojomo 2017-10-06
If it's in some specific `keras_utils`-like module, then exactly the Keras users who need it will find it – guided there by demo code like that in the notebook. 

Or, it just gets inlined or defined where needed. Or is just a documentation FAQ answer somewhere, to the question "How do I convert my KeyedVectors word-vectors into a Keras `Embedding`?"

It just seems awkward to me for `KeyedVectors` to have this trivial 1-liner that's strictly a keras-ism. It'd be better, for both modularity and avoiding the import-of-something-that-might-not-be-there-tap-dance, for keras-specific knowledge of how to turn a raw array like `syn0` into the keras type `Embedding` to be somewhere else. 

(And even if it does stay here, its name should probably include `keras` – say, `get_keras_embedding()` – to reflect its specificity here, not returning a more generic or local idea of an 'embedding'. )