@menshikh-iv 2017-10-06
Distinct submodule for one function - bad idea (especially considering that I'm currently doing refactoring and simplifying gensim).
About the name - I completely agree, this is missing information that **must** figured in a method name.